User Login
one issue is that cookie does not expire immediately after browser is closed. Googled that and find out that it depends on browser's settings. It does expire in chrome, but not straight away. Also, I do not know how to pass data between controllers or pass it straight to the view. Everything else should be working fine